finals/Johnnie Walker/requirements.txt
Flask==1.1.2
pyjwt==1.0.0
pyotp==2.4.0
